refactor: Move event primers into activity subfolders.
