+++ libopenage/gamestate/activity/event/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_sources(libopenage
+    command_in_queue.cpp
+    wait.cpp
+)
