+++ libopenage/gamestate/event/wait.h
@@ -45,23 +45,6 @@ public:
 	                                 const time::time_t &at) override;
 };
 
-/**
- * Primer for wait events in the activity system.
- *
- * @param time Wait until this time. If the time is in the past, the event is executed immediately.
- * @param entity Game entity.
- * @param loop Event loop that the event is registered on.
- * @param state Game state.
- * @param next_id ID of the next node in the activity graph.
- *
- * @return Scheduled event.
- */
-std::shared_ptr<openage::event::Event> primer_wait(const time::time_t &time,
-                                                   const std::shared_ptr<GameEntity> &entity,
-                                                   const std::shared_ptr<openage::event::EventLoop> &loop,
-                                                   const std::shared_ptr<gamestate::GameState> &state,
-                                                   size_t next_id);
-
 } // namespace event
 } // namespace gamestate
 } // namespace openage
